async_exec() use a pty instead of pipes / async_write() handle list arg
The current implementation provides good api and seams well integrated.
However, it suffered from buffering issues due to the use of pipes to
communicate between vim and child processes.

In order to address this issue, this patch use a pseudo-terminal between
vim and child processes.

  -------          -------------------           ---------
  |     | -------> | Master | Slave  | --------> | Child |
  | Vim |          |        |        |           |       |
  |     | <------- |  Pty   |  Pty   | <-------- | Proc. |
  -------          -------------------           ---------

Brief explanation:

When using pipes and pty IPCs, depending on which system calls we use,
we may use 3 buffers, 2 within libc and 1 within the kernel.
Moreover, there are 3 types of buffering: unbeffered,
line buffered (up to a new line character), block buffered (fixed size).

Pipe's default buffer is block buffering, which were the cause of the
latence if there were not enough data's written by child processes.

In order for vim to be responsive enough for an asynchronous exchanges
with a child process, we should setup at least a line buffering, which
is the default mode for ptys.

Here are few ideas which poped up when analysing this solution:

- pipes:
  - kernel pipes buffers are block buffered, their size are immutable
    (without re-compiling)
  - processes (like shells) can reset predefined setup - and/or - expect
    to interact with a terminal (e.g. unable to use setbuf(), undefined
    behavior)

- pty:
  (certainly the cleanest way)
  - line buffering is the default mode
  - straightforward and only few changes to adapt the current implementation
  - won't work on something else than *nix (read below)

- unix socket:
  - buffer's size are tunable
  - line buffering behavior is better, thus dealing between required
    interactivity and buffer size may be tricky.

- ip socket:
  - does it make sense to implement network in vim?
  - force the use of ip's aware processes to communicate with

Regarding this implementation:

  - thanks to the expect's team (http://expect.nist.gov/) and their work on
    unbuffer, the same robust method have been used to synchronize parent
    and childs, as weel as retreiving child status.
  - it is possible to use forkpty() (on some systems), but a "made by hand"
    solution was adopted instead.
  - we do not distinguished between stdout and stderr (yet?)

Thanks to B. Trojanowski and M. Weber for the base implementation.

Other OS:

  If other OSes enable the use of vim's clientserver X protocol, it is
  still possible to use M.Weber's vim-addon-async.
  The native method is mostly useful for *nix terminal anyway.

TODO:
  - look after the error when calling async_exec twice with the same ctx
    (even if the process has terminatedit)
  - clean ctx.pid when terminated?
  - vim-addon-async tests almost pass. It all works (raw checked have
    been done) but #Binary() still have an issue.
  - Is it the best way to handle "\n"?
  - ensime is starting fine :)